Change audio back into a script
This changes audio back to a mere script from a module, by removing the
import statement and adds a global howler namespace.

This is still not ideal, as the return values of those functions still
show up as `any` instead of `howler.Howl` or `Howl` for consuming
libraries. This needs to be improved in the future.