Add generic parameters to MouseInteractionManager
Adds generic parameters that correspond to the object the that the
interaction manager manages the mouse for and the target of interaction
events.

Adjusts the typing of `PlaceableObject.mouseInteractionManager` so that
the correct object and target is inferred. Creates an overload on the
controlIcon parameter of `null` for PlaceableObject classes that do not
have a ControlIcon such as Token.

Overrides the typing of `Wall.mouseInteractionManager` to handle its
special case of using the endpoints as the interaction target.

Adds tests to verify that the type of
`PlaceableObject.mouseInteractionManager` is correctly determined for
each version. Adds a test for the specific case of `Wall`.