Fix most type errors
The change to `TokenConfig.FormData` is that it inherits from its
parent. We add the `extends AnyObject` since that's what its parent
returns. The more robust way would likely be to define a `FormData` type
in the `DocumentSheet` namespace, and reference that instead. But,
that's a slightly bigger change than fixing things that are broken.

The change to `ShaderField.Options` is that `DataField.Options` doesn't
exist, but `DataFieldOptions` does.

The change to `BasePackage#name` is that it doesn't seem to be inferring
the correct type from the `this` context. Since the `name` comes from
whatever the `PackageSchema` is, we use that instead.